refactor: move offerCow to its own file
