+++ src/game-logic/reducers/index.js
@@ -75,17 +75,7 @@ export * from './purchaseSmelter'
 export * from './purchaseCowPen'
 export * from './purchaseStorageExpansion'
 export * from './hugCow'
-
-/**
- * @param {farmhand.state} state
- * @param {string} cowId
- * @returns {farmhand.state}
- */
-export const offerCow = (state, cowId) => {
-  state = { ...state, cowIdOfferedForTrade: cowId }
-
-  return state
-}
+export * from './offerCow'
 
 /**
  * @param {farmhand.state} state
